requirements.txt
Django==3.1.2
gunicorn==20.0.4
djangorestframework==3.12.1
